requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.2.2
Django==5.0.4
django-filter==24.2
flake8==7.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mccabe==0.7.0
pillow==10.3.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-telegram-bot==2.4
sniffio==1.3.1
sqlparse==0.5.0
tzdata==2024.1
